TaskQueue should remove all tasks when requested to remove tasks up to a size bigger than the total task size in the queue
PiperOrigin-RevId: 614672257

+++ tensorflow/core/kernels/batching_util/batch_scheduler.h
@@ -90,16 +90,16 @@ class TaskQueue {
   // Removes tasks from the front of the queue as many as possible as long as
   // the sum of sizes of the removed tasks don't exceed the 'size' given as the
   // argument.
-  std::vector<std::unique_ptr<TaskType>> RemoveTask(size_t size);
+  std::vector<std::unique_ptr<TaskType>> RemoveTask(int size);
 
   // Returns true iff the queue contains 0 tasks.
   bool empty() const;
 
   // Returns the number of tasks in the queue.
-  size_t num_tasks() const;
+  int num_tasks() const;
 
   // Returns the sum of the task sizes.
-  size_t size() const;
+  int size() const;
 
  private:
   mutable mutex mu_;
@@ -108,7 +108,7 @@ class TaskQueue {
   std::deque<std::unique_ptr<TaskType>> tasks_ TF_GUARDED_BY(mu_);
 
   // The sum of the sizes of the tasks in 'tasks_'.
-  size_t size_ TF_GUARDED_BY(mu_) = 0;
+  int size_ TF_GUARDED_BY(mu_) = 0;
 
   // Whether the queue is empty.
   std::atomic<bool> empty_ TF_GUARDED_BY(mu_){true};
@@ -147,18 +147,19 @@ std::unique_ptr<TaskType> TaskQueue<TaskType>::RemoveTask() {
 
 template <typename TaskType>
 std::vector<std::unique_ptr<TaskType>> TaskQueue<TaskType>::RemoveTask(
-    size_t size) {
+    int size) {
   {
     mutex_lock l(mu_);
     if (tasks_.empty()) {
       return {};
     }
 
-    size_t size_lower_bound = size_ - size;
+    int size_lower_bound = size_ - size;
     std::vector<std::unique_ptr<TaskType>> remove_tasks;
     while (!tasks_.empty() &&
-           size_ - tasks_.front()->size() >= size_lower_bound) {
-      size_ -= tasks_.front()->size();
+           size_ - static_cast<int>(tasks_.front()->size()) >=
+               size_lower_bound) {
+      size_ -= static_cast<int>(tasks_.front()->size());
       remove_tasks.push_back(std::move(tasks_.front()));
       tasks_.pop_front();
       if (tasks_.empty()) {
@@ -178,7 +179,7 @@ bool TaskQueue<TaskType>::empty() const {
 }
 
 template <typename TaskType>
-size_t TaskQueue<TaskType>::num_tasks() const {
+int TaskQueue<TaskType>::num_tasks() const {
   {
     mutex_lock l(mu_);
     return tasks_.size();
@@ -186,7 +187,7 @@ size_t TaskQueue<TaskType>::num_tasks() const {
 }
 
 template <typename TaskType>
-size_t TaskQueue<TaskType>::size() const {
+int TaskQueue<TaskType>::size() const {
   {
     mutex_lock l(mu_);
     return size_;
